requirements.txt
flask
google-generativeai
supabase
